Update hamcrest/library/number/iscloseto.py
The existing test for "isnumeric" was very non-pythonic, passing only if "value" was an instance of int/float/long. This is clearly wrong ;)  The tests would fail incorrectly for (amongst others) perfectly reasonable numpy scalars and decimal values, etc. Better to duck-check; if an object doesn't throw on math.fabs (which we'll be using in the real test) and it supports basic math operators, then it should be considered numeric for the purposes of this test.