Delete str() method that causing UnicodeEncodeError
If 'matcher' mismatching with 'actual' and one of them has non-ascii encoding, then description has the same type. When str() method trying encode 'description' with ASCII codec it causes UnicodeEncodeError. To avoid getting this error and do not worry about converting to a different encoding, it can simply display the string as is.