add has_return_value matcher.
This new matcher will match against the return value of an object's method.
In the world of web automation, this matcher will be particularly useful to
test for things like whether an element is displayed by matching against its
`is_displayed()` method return value, or checking class strings, or any number
of things an automated tester might need to check.