Refactor the docs to promote user functions
The functions that users most often use are the matcher functions,
however the docs show the internal matcher classes first. This means
that users have to scan through a bunch of internal detail to get to the
documentation they're most likely interested in. This change switches
things around to just document the actual matcher functions.